# swagger.json
{
...
  "info": {
    "title": "API Documentation",
    "version": "0.0.1"
  },
  "paths": {
...
    "/admin/endpoint": {
      "post": {
        "parameters": [
          {
            "description": "Admin API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          },
          {
            "description": "A valid token for verification",
            "in": "query",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "active": {
                  "description": "Whether the endpoint is active (default: true)",
                  "type": "boolean"
                },
                "cost": {
                  "description": "Cost in balance units for each call to this endpoint (default: 1.0, can be 0 or fractional values like 0.5)",
                  "format": "float",
                  "type": "number"
                },
                "description": {
                  "description": "Optional description of the endpoint",
                  "type": "string"
                },
                "endpoint_name": {
                  "description": "A user-friendly name for the endpoint",
                  "type": "string"
                },
                "endpoint_path": {
                  "description": "The API endpoint path (e.g., /llm/custom)",
                  "type": "string"
                }
              },
              "required": [
                "endpoint_path",
                "endpoint_name"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Endpoint created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication error"
          },
          "403": {
            "description": "Forbidden - not an admin"
          },
          "409": {
            "description": "Conflict - endpoint already exists"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Add a new endpoint to the endpoints table (Admin only endpoint)",
        "tags": [
          "Admin Functions"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "Admin API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          },
          {
            "description": "A valid token for verification",
            "in": "query",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "active": {
                  "description": "New active status (optional)",
                  "type": "boolean"
                },
                "cost": {
                  "description": "New cost in balance units for each call (optional)",
                  "type": "integer"
                },
                "description": {
                  "description": "New description (optional)",
                  "type": "string"
                },
                "endpoint_id": {
                  "description": "ID of the endpoint to update",
                  "type": "string"
                },
                "endpoint_name": {
                  "description": "New user-friendly name (optional)",
                  "type": "string"
                },
                "endpoint_path": {
                  "description": "New API endpoint path (optional)",
                  "type": "string"
                }
              },
              "required": [
                "endpoint_id"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Endpoint updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication error"
          },
          "403": {
            "description": "Forbidden - not an admin"
          },
          "404": {
            "description": "Endpoint not found"
          },
          "409": {
            "description": "Conflict - endpoint path already exists"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Update an existing endpoint in the endpoints table (Admin only endpoint)",
        "tags": [
          "Admin Functions"
        ]
      }
    },
    "/admin/endpoints": {
      "get": {
        "parameters": [
          {
            "description": "Admin API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          },
          {
            "description": "A valid token for verification",
            "in": "query",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "description": "Filter by active status (optional)",
            "in": "query",
            "name": "active",
            "required": false,
            "type": "boolean"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Endpoints retrieved successfully"
          },
          "401": {
            "description": "Authentication error"
          },
          "403": {
            "description": "Forbidden - not an admin"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Get all endpoints from the endpoints table (Admin only endpoint)",
        "tags": [
          "Admin Functions"
        ]
      }
    },
...
    "/token": {
      "get": {
        "parameters": [
          {
            "description": "API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token generated successfully",
            "schema": {
              "properties": {
                "access_token": {
                  "description": "generated access token to use with api calls",
                  "type": "string"
                },
                "expires_in": {
                  "description": "Time in seconds until token expiration",
                  "format": "seconds",
                  "type": "integer"
                },
                "expires_on": {
                  "description": "Token expiration timestamp",
                  "format": "date-time",
                  "type": "string"
                },
                "token_type": {
                  "description": "Type of token generated",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "properties": {
                "error": {
                  "example": "Authentication Error",
                  "type": "string"
                },
                "message": {
                  "example": "Missing API Key header (API-Key) or Invalid API Key",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "properties": {
                "error": {
                  "example": "Server Error",
                  "type": "string"
                },
                "message": {
                  "example": "Error generating token",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Generate a token for API access using a valid api authentication key.",
        "tags": [
          "Token Service"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "API Key for authentication. Add your token with the X-Token header.",
      "in": "header",
      "name": "X-Token",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}